Add files via upload
update/ now working like Bra & Emi with opp Live (fixed do_op count  started now (cycles_to_do - 2))
FWSIO

+++ play_game.c
@@ -6,7 +6,7 @@
 /*   By: lgeorgia <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/05 19:18:36 by lgeorgia          #+#    #+#             */
-/*   Updated: 2019/11/11 20:18:22 by lgeorgia         ###   ########.fr       */
+/*   Updated: 2019/11/12 20:17:25 by lgeorgia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,13 +48,11 @@ int	ft_check_live_crg(t_war *war)
 		{
 			help->die = 1;
 		}
-		else if (help->die == 0)
+		if (help->die == 0)
 		{
 			help->live = 0;
 			war->numb_crg++;
 		}
-		if (help->next == war->top)
-			return (war->numb_crg);
 		help = help->next;
 	}
 	return (war->numb_crg);
@@ -70,17 +68,27 @@ void	ft_play_game(t_war *war)
 	{
 		help = war->top;
 		// printf("\ncycles = %d\n", war->cycle);
+		if (war->to_die <= 0)
+		{
+				ft_putstr("\nWinner is player number ");
+				ft_putnbr(war->winner);
+				write(1, "\n", 1);
+				printf("CYCLE MAX = %d DIE<<<<<= 0\n", war->all_cycle);
+				exit(1);
+		}
 		if (war->cycle == war->to_die)
 		{
+			printf("\ncalled live = %d\n", war->live);
+			war->check_num++;
 			if (ft_check_live_crg(war) == 0) // someone alive and kill no life crg
 			{
 			
 				ft_putstr("\nWinner is player number ");
 				ft_putnbr(war->winner);
 				write(1, "\n", 1);
+				printf("CYCLE MAX = %d war_to_die = %d\n", war->all_cycle, war->to_die);
 				exit(1);
 			}
-			war->check_num++;
 			if (war->check_num == 10 || war->live >= NBR_LIVE)
 			{
 
@@ -88,11 +96,12 @@ void	ft_play_game(t_war *war)
 				war->check_num = 0;
 			}
 			war->cycle = 0;
+			war->live = 0;
 		}
-		else
-		{
-			ft_check_status_of_crg(war, help);
-			war->cycle++;
-		}
+		ft_check_status_of_crg(war, help);
+		war->cycle++;
+		war->all_cycle++;
+		// printf("MAX CYCLE = %d CYCLE_TO_DIE = %d\n", war->all_cycle, war->to_die);
+		// printf("pos 1 pl = %d pos 2 pl = %d\n", war->bot->pos, war->top->pos);
 	}
 }
